moving navigation to app,js
