+++ src/funktionality/navigation/movePointer.js
