+++ src/funktionality/navigation/navigateFight.js
